requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pillow>=10.0.0
patool>=1.12
python-magic>=0.4.27
google-generativeai>=0.3.3  # Must match documentation version
anthropic>=0.25.0
openai>=1.12.0
httpx>=0.27.0